refactor: VerificationService.verifyPassword 삭제

+++ src/main/java/com/communify/domain/member/application/PasswordBasedMemberWithdrawService.java
@@ -1,9 +1,12 @@
 package com.communify.domain.member.application;
 
-import com.communify.domain.auth.application.VerificationService;
 import com.communify.domain.auth.application.LoginService;
+import com.communify.domain.auth.error.exception.InvalidPasswordException;
 import com.communify.domain.member.dao.MemberRepository;
+import com.communify.domain.member.dto.MemberInfo;
 import com.communify.domain.member.dto.MemberWithdrawRequest;
+import com.communify.domain.member.error.exception.MemberNotFoundException;
+import com.communify.global.util.PasswordEncryptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -11,15 +14,21 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class PasswordBasedMemberWithdrawService implements MemberWithdrawService {
 
+    private final MemberFindService memberFindService;
     private final MemberRepository memberRepository;
-    private final VerificationService verificationService;
     private final LoginService loginService;
 
     @Override
     public void withdraw(MemberWithdrawRequest request, Long memberId) {
-        String password = request.getPassword();
+        MemberInfo memberInfo = memberFindService.findMemberInfoById(memberId)
+                .orElseThrow(() -> new MemberNotFoundException(memberId));
 
-        verificationService.verifyPassword(password, memberId);
+        String password = request.getPassword();
+        String hashed = memberInfo.getHashed();
+
+        if (!PasswordEncryptor.isMatch(password, hashed)) {
+            throw new InvalidPasswordException(password);
+        }
 
         memberRepository.deleteById(memberId);
         loginService.logout();
